Move IOHandler to separate project
