+++ DelfinForWindows.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31105.61
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DelfinForWindows", "DelfinForWindows\DelfinForWindows.csproj", "{F8322398-99D3-4009-9F05-CC7AA5AC4575}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DelfinForWindows", "DelfinForWindows\DelfinForWindows.csproj", "{F8322398-99D3-4009-9F05-CC7AA5AC4575}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IOHandler", "IOHandler\IOHandler.csproj", "{5C0DAB81-5C1B-4290-9668-201B6C8B0A91}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{F8322398-99D3-4009-9F05-CC7AA5AC4575}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F8322398-99D3-4009-9F05-CC7AA5AC4575}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F8322398-99D3-4009-9F05-CC7AA5AC4575}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5C0DAB81-5C1B-4290-9668-201B6C8B0A91}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5C0DAB81-5C1B-4290-9668-201B6C8B0A91}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5C0DAB81-5C1B-4290-9668-201B6C8B0A91}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5C0DAB81-5C1B-4290-9668-201B6C8B0A91}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
